Rewrite simply supported equal symmetric tests
